feat: ViewController의 기능 분리하여 DailyBoxOfficeCollectionView 구현
